+++ src/main/java/com/foxminded/universitytimetable/dao/StatisticsDAO.java
@@ -3,7 +3,7 @@ package com.foxminded.universitytimetable.dao;
 import java.util.Date;
 
 public interface StatisticsDAO {
-    int getProfessorEmployment(int professorId, Date from, Date till);
+    long getProfessorEmployment(int professorId, Date from, Date till);
 
-    int getGroupEmployment(int groupId, Date from, Date till);
+    long getGroupEmployment(int groupId, Date from, Date till);
 }
